move factory to separate file
